Cleanup warnings
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@1193 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ addon/ms-logon/MSLogonACL/vncImportACL.cpp
@@ -447,7 +447,7 @@ bool vncScanInput::Tokenize(const TCHAR *line, TCHAR **token){
 			line++; // Eat closing quote
 			break;
 		default:
-			len = GetWordLength(line);
+			len = (int)GetWordLength(line);
 			tokenCount = AddToken(token, tokenCount, &line, len);
 			break;
 		}
@@ -482,7 +482,7 @@ int vncScanInput::GetQuoteLength(const TCHAR *line){
     size_t len = 0;
 	if (eoq)
 		len = eoq - line;
-	return len;
+	return (int)len;
 }
 
 int vncScanInput::AddToken(TCHAR **token, int tokenCount, const TCHAR **line, int len){
